requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
langchain-core
langchain-google-genai
google-genai
groq
duckdb
matplotlib
plotly
pydantic
pydantic-settings
python-dotenv
websockets
python-multipart
aiofiles
RestrictedPython
